The export behaviour for the XBEL export function (implemented in accessories/mm2xbel.xsl) could be improved. Currently only those nodes are exported, which are links or leaves. If the idea was to only capture parents of links, then this is still suboptimal, since there could certainly be nodes, which have non-link descendants.
I want to preserve all information in the original FreeMind map structure
(I am importing it with Treeline, an outliner which can import the XBEL
format), so I want to export all non-link nodes as folders.

So that's what the patch does. It modifies the currently distributed
mm2xbel.xls file so that all nodes are exported- if it contains a link, as
a link, and if it doesn't- as an XBEL folder.

Maybe it is also worth preserving a minimal structure of links with only
their parents (with a parameter or another xsl), but that's not what the
current implementation does.

All the best for this totally cool tool!
Vassil